Intial modifications for N bringup
-- disable cmte, for now
-- adapt to new API in a few spots
-- move last app anims to SystemUI, less invasive
-- KillApp action: Show package or process label
   in toast string. Blacklist a few new gapps
   packages that show up in foreground
-- Add Split screen action
-- Properly decode and set density on icons from storage

+++ src/com/android/internal/utils/du/ActionHandler.java
@@ -33,6 +33,7 @@ import android.content.ContentResolver;
 import android.content.Context;
 import android.content.ComponentName;
 import android.content.Intent;
+import android.content.pm.ApplicationInfo;
 import android.content.pm.PackageManager;
 import android.content.pm.ResolveInfo;
 import android.content.res.Resources;
@@ -125,6 +126,7 @@ public class ActionHandler {
     public static final String SYSTEMUI_TASK_CLEAR_NOTIFICATIONS = "task_clear_notifications";
     public static final String SYSTEMUI_TASK_VOLUME_PANEL = "task_volume_panel";
     public static final String SYSTEMUI_TASK_EDITING_SMARTBAR = "task_editing_smartbar";
+    public static final String SYSTEMUI_TASK_SPLIT_SCREEN = "task_split_screen";
 
     public static final String INTENT_SHOW_POWER_MENU = "action_handler_show_power_menu";
     public static final String INTENT_TOGGLE_SCREENRECORD = "action_handler_toggle_screenrecord";
@@ -162,7 +164,8 @@ public class ActionHandler {
         ImeArrowUp(SYSTEMUI_TASK_IME_NAVIGATION_UP, SYSTEMUI, "label_action_ime_up", "ic_sysbar_ime_up"),
         ClearNotifications(SYSTEMUI_TASK_CLEAR_NOTIFICATIONS, SYSTEMUI, "label_action_clear_notifications", "ic_sysbar_clear_notifications"),
         VolumePanel(SYSTEMUI_TASK_VOLUME_PANEL, SYSTEMUI, "label_action_volume_panel", "ic_sysbar_volume_panel"),
-        EditingSmartbar(SYSTEMUI_TASK_EDITING_SMARTBAR, SYSTEMUI, "label_action_editing_smartbar", "ic_sysbar_editing_smartbar");
+        EditingSmartbar(SYSTEMUI_TASK_EDITING_SMARTBAR, SYSTEMUI, "label_action_editing_smartbar", "ic_sysbar_editing_smartbar"),
+        SplitScreen(SYSTEMUI_TASK_SPLIT_SCREEN, SYSTEMUI, "label_action_split_screen", "ic_sysbar_docked");
 
         String mAction;
         String mResPackage;
@@ -200,7 +203,7 @@ public class ActionHandler {
             SystemAction.ImeArrowLeft, SystemAction.ImeArrowRight,
             SystemAction.ImeArrowUp, SystemAction.InAppSearch,
             SystemAction.VolumePanel, SystemAction.ClearNotifications,
-            SystemAction.EditingSmartbar
+            SystemAction.EditingSmartbar, SystemAction.SplitScreen
     };
 
     public static class ActionIconResources {
@@ -261,6 +264,13 @@ public class ActionHandler {
             } else if (TextUtils.equals(action, SYSTEMUI_TASK_SCREENRECORD)) {
                 if (!DUActionUtils.getBoolean(context, "config_enableScreenrecordChord",
                         DUActionUtils.PACKAGE_ANDROID)) {
+                    continue;
+                }
+            } else if (TextUtils.equals(action, SYSTEMUI_TASK_EDITING_SMARTBAR)) {
+                // don't allow smartbar editor on Fling
+                if (Settings.Secure.getIntForUser(context.getContentResolver(),
+                        Settings.Secure.NAVIGATION_BAR_MODE, 0,
+                        UserHandle.USER_CURRENT) == 1) {
                     continue;
                 }
             }
@@ -341,7 +351,7 @@ public class ActionHandler {
             IStatusBarService service = getStatusBarService();
             if (service != null) {
                 try {
-                    service.expandSettingsPanel();
+                    service.expandSettingsPanel(null);
                 } catch (RemoteException e) {
                 }
             }
@@ -357,6 +367,16 @@ public class ActionHandler {
             }
         }
 
+        private static void splitScreen() {
+            IStatusBarService service = getStatusBarService();
+            if (service != null) {
+                try {
+                    service.toggleSplitScreen();
+                } catch (RemoteException e) {
+                }
+            }
+        }
+/*
         private static void fireIntentAfterKeyguard(Intent intent) {
             IStatusBarService service = getStatusBarService();
             if (service != null) {
@@ -366,7 +386,7 @@ public class ActionHandler {
                 }
             }
         }
-
+*/
         private static void clearAllNotifications() {
             IStatusBarService service = getStatusBarService();
             if (service != null) {
@@ -389,7 +409,7 @@ public class ActionHandler {
     public static void preloadRecentApps() {
         StatusBarHelper.preloadRecentApps();
     }
-
+/*
     public static void performTaskFromKeyguard(Context ctx, String action) {
         // null: throw it out
         if (action == null) {
@@ -406,7 +426,7 @@ public class ActionHandler {
             performTask(ctx, action);
         }
     }
-
+*/
     public static void performTask(Context context, String action) {
         // null: throw it out
         if (action == null) {
@@ -592,6 +612,9 @@ public class ActionHandler {
         } else if (action.equals(SYSTEMUI_TASK_EDITING_SMARTBAR)) {
             editingSmartbar(context);
             return;
+        } else if (action.equals(SYSTEMUI_TASK_SPLIT_SCREEN)) {
+            StatusBarHelper.splitScreen();
+            return;
         }
     }
 
@@ -623,8 +646,8 @@ public class ActionHandler {
 
         if (lastTask != null) {
             final ActivityOptions opts = ActivityOptions.makeCustomAnimation(context,
-                    DUActionUtils.getIdentifier(context, "last_app_in", "anim", DUActionUtils.PACKAGE_ANDROID),
-                    DUActionUtils.getIdentifier(context, "last_app_out", "anim", DUActionUtils.PACKAGE_ANDROID));
+                    DUActionUtils.getIdentifier(context, "last_app_in", "anim", DUActionUtils.PACKAGE_SYSTEMUI),
+                    DUActionUtils.getIdentifier(context, "last_app_out", "anim", DUActionUtils.PACKAGE_SYSTEMUI));
             am.moveTaskToFront(lastTask.id, ActivityManager.MOVE_TASK_NO_USER_ACTION,
                     opts.toBundle());
         }
@@ -844,7 +867,7 @@ public class ActionHandler {
                     defaultHomePackage = res.activityInfo.packageName;
                 }
                 IActivityManager am = ActivityManagerNative.getDefault();
-                boolean targetKilled = false;
+                String pkgName;
                 List<RunningAppProcessInfo> apps = am.getRunningAppProcesses();
                 for (RunningAppProcessInfo appInfo : apps) {
                     int uid = appInfo.uid;
@@ -859,21 +882,36 @@ public class ActionHandler {
                                 if (!pkg.equals("com.android.systemui")
                                         && !pkg.equals(defaultHomePackage)
                                         && !pkg.equals("com.google.android.googlequicksearchbox")
+                                        && !pkg.equals("com.google.android.gms")
+                                        && !pkg.equals("com.google.android.ext.services")
+                                        && !pkg.equals("android.ext.services")
                                         && !isPackageLiveWalls(context, pkg)) {
+                                    try {
+                                        ApplicationInfo applicationInfo = context
+                                                .getPackageManager().getPackageInfo(pkg, 0).applicationInfo;
+                                        pkgName = applicationInfo.loadLabel(
+                                                context.getPackageManager()).toString();
+                                    } catch (Exception e) {
+                                        // cheat just a little, highly unlikely event
+                                        pkgName = "App";
+                                    }
                                     am.forceStopPackage(pkg,
                                             UserHandle.USER_CURRENT);
-                                    targetKilled = true;
-                                    break;
+                                    Resources systemUIRes = DUActionUtils.getResourcesForPackage(context, DUActionUtils.PACKAGE_SYSTEMUI);
+                                    int ident = systemUIRes.getIdentifier("app_killed_message", DUActionUtils.STRING, DUActionUtils.PACKAGE_SYSTEMUI);
+                                    String toastMsg = systemUIRes.getString(ident, pkgName);
+                                    Toast.makeText(context, toastMsg, Toast.LENGTH_SHORT).show();
+                                    return;
                                 }
                             }
                         } else {
+                            pkgName = appInfo.processName;
                             Process.killProcess(appInfo.pid);
-                            targetKilled = true;
-                        }
-                        if (targetKilled) {
-                            Toast.makeText(context, com.android.internal.R.string.app_killed_message,
-                                    Toast.LENGTH_SHORT).show();
-                            break;
+                            Resources systemUIRes = DUActionUtils.getResourcesForPackage(context, DUActionUtils.PACKAGE_SYSTEMUI);
+                            int ident = systemUIRes.getIdentifier("app_killed_message", DUActionUtils.STRING, DUActionUtils.PACKAGE_SYSTEMUI);
+                            String toastMsg = systemUIRes.getString(ident, pkgName);
+                            Toast.makeText(context, toastMsg, Toast.LENGTH_SHORT).show();
+                            return;
                         }
                     }
                 }
@@ -929,7 +967,8 @@ public class ActionHandler {
 
     public static void turnOffLockTask() {
         try {
-            ActivityManagerNative.getDefault().stopLockTaskModeOnCurrent();
+//            ActivityManagerNative.getDefault().stopLockTaskModeOnCurrent();
+        	ActivityManagerNative.getDefault().stopLockTaskMode();
         } catch (Exception e) {
         }
     }
